+++ unoplat-code-confluence-docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.32.1](https://github.com/unoplat/unoplat-code-confluence/compare/unoplat-code-confluence-docs-v0.32.0...unoplat-code-confluence-docs-v0.32.1) (2026-02-14)
+
+
+### Bug Fixes
+
+* **discord-links:** discord links were broken across docs/app/readme ([ad29fa5](https://github.com/unoplat/unoplat-code-confluence/commit/ad29fa5577d2b0f1555304f294896e7a9fd5fcc4))
+* **discord-links:** discord links were broken across docs/app/readme ([43fd793](https://github.com/unoplat/unoplat-code-confluence/commit/43fd7936dd1940b6a4dfe4a2838d252e6ec205ad))
+
 ## [0.32.0](https://github.com/unoplat/unoplat-code-confluence/compare/unoplat-code-confluence-docs-v0.31.0...unoplat-code-confluence-docs-v0.32.0) (2026-02-11)